+++ CHANGELOG.md
@@ -6,5 +6,6 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Initial release of the library.
+- Added base rules from iserv-dev. #42732
+- Initial release of the library. #42696